Выполнить второй этап дипломного проекта

+++ stats/service/src/main/java/ru/practicum/ewm/model/StatsProjection.java
@@ -5,5 +5,5 @@ public interface StatsProjection {
 
     String getUri();
 
-    Long getHits();
+    Integer getHits();
 }
